Split verification functions into its own service.
